reuse existing near chain file
